A bit a cleanup come due.
