--- experiments/billable/src/app/shared/invoice.ts
@@ -1,18 +1,0 @@
-import { InvoiceItem, InvoiceTaxes } from "../services/invoices";
-
-
-export function calculateSubtotal(items: InvoiceItem[]) {
-  let sum = 0;
-  for (const item of items) {
-    sum += item.quantity * item.price;
-  }
-  return sum;
-}
-
-export function calculateTaxes(subtotal: number, taxes: InvoiceTaxes[]) {
-  let sum = 0;
-  for (const tax of taxes) {
-    sum += subtotal * tax.amount;
-  }
-  return sum;
-}
